requirements.txt
streamlit
torch
torchvision
Pillow
matplotlib
numpy
